backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
SpeechRecognition==3.10.1
gTTS==2.5.1
python-dotenv==1.0.1
pyspellchecker==0.7.2
google-generativeai==0.3.2
python-dotenv==1.0.1

